requirements.txt
fastapi
uvicorn
pyyaml
pydantic
scikit-learn
transformers
torch
